fix: formatting and lint errors
