+++ src/components/ThemeProvider.tsx
@@ -6,5 +6,5 @@ import { ThemeProvider as NextThemesProvider } from "next-themes";
 type ThemeProviderProps = React.ComponentProps<typeof NextThemesProvider>;
 
 export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
-    return <NextThemesProvider {...props}>{children}</NextThemesProvider>;
+  return <NextThemesProvider {...props}>{children}</NextThemesProvider>;
 }
