+++ src/components/landing/AboutSection.tsx
@@ -29,7 +29,10 @@ const features = [
 
 export function AboutSection() {
   return (
-    <section id="about" className="py-24 bg-surface-raised relative overflow-hidden">
+    <section
+      id="about"
+      className="py-24 bg-surface-raised relative overflow-hidden"
+    >
       <div className="absolute inset-0 pointer-events-none" />
 
       <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 relative">
